@@ -1,18 +1,21 @@
 using System;
 using System.Collections.Generic;
 
-namespace NCoreUtils.Google
+namespace NCoreUtils.Google;
+
+public readonly struct GoogleObjectsPage
 {
-    public class GoogleObjectsPage
+    private readonly IReadOnlyList<string>? _prefixes;
+
+    private readonly IReadOnlyList<GoogleObjectData>? _items;
+
+    public IReadOnlyList<string> Prefixes => _prefixes ?? Array.Empty<string>();
+
+    public IReadOnlyList<GoogleObjectData> Items => _items ?? Array.Empty<GoogleObjectData>();
+
+    public GoogleObjectsPage(IReadOnlyList<string>? prefixes, IReadOnlyList<GoogleObjectData>? items)
     {
-        public IReadOnlyList<string> Prefixes { get; }
-
-        public IReadOnlyList<GoogleObjectData> Items { get; }
-
-        public GoogleObjectsPage(IReadOnlyList<string>? prefixes, IReadOnlyList<GoogleObjectData>? items)
-        {
-            Prefixes = prefixes ?? Array.Empty<string>();
-            Items = items ?? Array.Empty<GoogleObjectData>();
-        }
+        _prefixes = prefixes;
+        _items = items;
     }
 }
