@@ -9,196 +9,210 @@ using System.Threading;
 using System.Threading.Tasks;
 using NCoreUtils.Google;
 
-namespace NCoreUtils
+namespace NCoreUtils;
+
+public sealed partial class GoogleCloudStorageUploader : IDisposable, IAsyncDisposable
 {
-    public sealed class GoogleCloudStorageUploader : IDisposable, IAsyncDisposable
+    private static MediaTypeHeaderValue? _applicationOctetStream;
+
+    private static MediaTypeHeaderValue ApplicationOctetStream
+        => _applicationOctetStream ??= MediaTypeHeaderValue.Parse("application/octet-stream");
+
+    public static int MinChunkSize { get; } = 256 * 1024;
+
+    public static int DefaultChunkSize { get; } = 512 * 1024;
+
+    private readonly IMemoryOwner<byte> _buffer;
+
+    private long _sent;
+
+    private long _isDisposed;
+
+    public event EventHandler<GoogleCloudStorageUploaderProgressArgs>? Progress;
+
+    private long Sent
     {
-        private sealed class SliceMemoryOwner : IMemoryOwner<byte>
+        get => _sent;
+        set
         {
-            private readonly IMemoryOwner<byte> _source;
-
-            private readonly int _size;
-
-            public SliceMemoryOwner(IMemoryOwner<byte> source, int size)
+            if (_sent != value)
             {
-                _source = source ?? throw new ArgumentNullException(nameof(source));
-                if (source.Memory.Length < size)
-                {
-                    throw new ArgumentException("Supplied memory is smaller than the requested slice size.", nameof(source));
-                }
-                _size = size;
-            }
-
-            public Memory<byte> Memory
-                => _source.Memory[.._size];
-
-            public void Dispose()
-                => _source.Dispose();
-        }
-
-        public static int MinChunkSize { get; } = 256 * 1024;
-
-        public static int DefaultChunkSize { get; } = 512 * 1024;
-
-        // readonly UnmanagedMemoryManager<byte> _buffer = new UnmanagedMemoryManager<byte>(512 * 1024);
-
-        private readonly IMemoryOwner<byte> _buffer;
-
-        private long _sent;
-
-        private long _isDisposed;
-
-        public event EventHandler<GoogleCloudStorageUploaderProgressArgs>? Progress;
-
-        private long Sent
-        {
-            get => _sent;
-            set
-            {
-                if (_sent != value)
-                {
-                    _sent = value;
-                    Progress?.Invoke(this, new GoogleCloudStorageUploaderProgressArgs(value));
-                }
+                _sent = value;
+                Progress?.Invoke(this, new GoogleCloudStorageUploaderProgressArgs(value));
             }
         }
+    }
 
-        public HttpClient Client { get; }
+    public HttpClient Client { get; }
 
-        public Uri EndPoint { get; }
+    public Uri EndPoint { get; }
 
-        public MediaTypeHeaderValue ContentType { get; }
+    public MediaTypeHeaderValue ContentType { get; }
 
-        public GoogleCloudStorageUploader(
-            HttpClient client,
-            Uri endpoint,
-            string? contentType = default,
-            IMemoryOwner<byte>? buffer = default)
+    public GoogleCloudStorageUploader(
+        HttpClient client,
+        Uri endpoint,
+        string? contentType = default,
+        IMemoryOwner<byte>? buffer = default)
+    {
+        Client = client ?? throw new ArgumentNullException(nameof(client));
+        EndPoint = endpoint ?? throw new ArgumentNullException(nameof(endpoint));
+        ContentType = string.IsNullOrEmpty(contentType)
+            ? ApplicationOctetStream
+            : MediaTypeHeaderValue.Parse(contentType);
+        if (buffer is null)
         {
-            Client = client ?? throw new ArgumentNullException(nameof(client));
-            EndPoint = endpoint ?? throw new ArgumentNullException(nameof(endpoint));
-            ContentType = MediaTypeHeaderValue.Parse(contentType ?? "application/octet-stream");
-            if (buffer is null)
+            var b = MemoryPool<byte>.Shared.Rent(DefaultChunkSize);
+            if (b.Memory.Length == DefaultChunkSize)
             {
-                var b = MemoryPool<byte>.Shared.Rent(DefaultChunkSize);
-                if (b.Memory.Length == DefaultChunkSize)
-                {
-                    _buffer = b;
-                }
-                else
-                {
-                    _buffer = new SliceMemoryOwner(b, DefaultChunkSize);
-                }
+                _buffer = b;
             }
             else
             {
-                var memoryLength = buffer.Memory.Length;
-                if (memoryLength < MinChunkSize)
-                {
-                    throw new InvalidOperationException($"Supplied memory is smaller than the minimum chunk size ({MinChunkSize}).");
-                }
-                var usableSize = (memoryLength / MinChunkSize) * MinChunkSize;
-                if (memoryLength == usableSize)
-                {
-                    _buffer = buffer;
-                }
-                else
-                {
-                    _buffer = new SliceMemoryOwner(buffer, usableSize);
-                }
+                _buffer = new SliceMemoryOwner(b, DefaultChunkSize);
             }
         }
-
-        public GoogleCloudStorageUploader(
-            HttpClient client,
-            string endpoint,
-            string? contentType = default,
-            IMemoryOwner<byte>? buffer = default)
-            : this(client, new Uri(endpoint, UriKind.Absolute), contentType, buffer)
-        { }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private void ThrowIfDisposed()
+        else
         {
-            if (0 != Interlocked.CompareExchange(ref _isDisposed, 0, 0))
+            var memoryLength = buffer.Memory.Length;
+            if (memoryLength < MinChunkSize)
             {
-                throw new ObjectDisposedException(nameof(GoogleCloudStorageUploader));
+                throw new InvalidOperationException($"Supplied memory is smaller than the minimum chunk size ({MinChunkSize}).");
             }
-        }
-
-        private async Task<HttpResponseMessage> SendChunkAsync(int size, bool final, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            var targetSize = Sent + size;
-            var content = new ReadOnlyMemoryContent(size == _buffer.Memory.Length ? _buffer.Memory : _buffer.Memory[..size]);
-            var headers = content.Headers;
-            headers.ContentLength = size;
-            headers.ContentType = ContentType;
-            headers.ContentRange = !final ? new ContentRangeHeaderValue(Sent, targetSize - 1L) : new ContentRangeHeaderValue(Sent, targetSize - 1L, targetSize);
-            using var request = new HttpRequestMessage(HttpMethod.Put, EndPoint) { Content = content };
-            return await Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, cancellationToken).ConfigureAwait(false);
-        }
-
-        async ValueTask<int> FillBuffer(Stream stream, CancellationToken cancellationToken)
-        {
-            var totalRead = 0;
-            int read;
-            do
+            var usableSize = memoryLength / MinChunkSize * MinChunkSize; // Floor to chunk size
+            if (memoryLength == usableSize)
             {
-                read = await stream.ReadAsync(_buffer.Memory[totalRead..], cancellationToken).ConfigureAwait(false);
-                totalRead += read;
-            }
-            while (read > 0 && totalRead < _buffer.Memory.Length);
-            return totalRead;
-        }
-
-        public async Task ConumeStreamAsync(Stream stream, CancellationToken cancellationToken)
-        {
-            var read = await FillBuffer(stream, cancellationToken).ConfigureAwait(false);
-            if (read == _buffer.Memory.Length)
-            {
-                // partial upload
-                var response = await SendChunkAsync(read, false, cancellationToken).ConfigureAwait(false);
-                if ((int)response.StatusCode == 308)
-                {
-                    // chunk upload successfull (TODO: check if response range is set properly)
-                    Sent += read;
-                    await ConumeStreamAsync(stream, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    throw new GoogleCloudStorageUploadException($"Upload chunk failed with status code {response.StatusCode}.");
-                }
+                _buffer = buffer;
             }
             else
             {
-                // FIXME: handle stream being divisable by chunk size
-                var response = await SendChunkAsync(read, true, cancellationToken).ConfigureAwait(false);
+                _buffer = new SliceMemoryOwner(buffer, usableSize);
+            }
+        }
+    }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public GoogleCloudStorageUploader(
+        HttpClient client,
+        string endpoint,
+        string? contentType = default,
+        IMemoryOwner<byte>? buffer = default)
+        : this(client, new Uri(endpoint, UriKind.Absolute), contentType, buffer)
+    { }
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    private void ThrowIfDisposed()
+    {
+        if (0 != Interlocked.CompareExchange(ref _isDisposed, 0, 0))
+        {
+            throw new ObjectDisposedException(nameof(GoogleCloudStorageUploader));
+        }
+    }
+
+    private async Task<HttpResponseMessage> SendChunkAsync(int size, bool final, CancellationToken cancellationToken)
+    {
+        cancellationToken.ThrowIfCancellationRequested();
+        var targetSize = Sent + size;
+        var content = new ReadOnlyMemoryContent(size == _buffer.Memory.Length ? _buffer.Memory : _buffer.Memory[..size]);
+        var headers = content.Headers;
+        headers.ContentLength = size;
+        headers.ContentType = ContentType;
+        headers.ContentRange = !final ? new ContentRangeHeaderValue(Sent, targetSize - 1L) : new ContentRangeHeaderValue(Sent, targetSize - 1L, targetSize);
+        using var request = new HttpRequestMessage(HttpMethod.Put, EndPoint) { Content = content };
+        return await Client.SendAsync(request, HttpCompletionOption.ResponseHeadersRead, cancellationToken).ConfigureAwait(false);
+    }
+
+    private async ValueTask<int> FillBuffer(Stream stream, CancellationToken cancellationToken)
+    {
+        var totalRead = 0;
+        int read;
+        do
+        {
+            read = await stream.ReadAsync(_buffer.Memory[totalRead..], cancellationToken).ConfigureAwait(false);
+            totalRead += read;
+        }
+        while (read > 0 && totalRead < _buffer.Memory.Length);
+        return totalRead;
+    }
+
+    public async Task UploadAsync(Stream stream, bool leaveOpen, CancellationToken cancellationToken)
+    {
+        await using var chunker = CreateChunkSource(stream, leaveOpen);
+        while (true)
+        {
+            var (size, final) = await chunker.FetchChunkAsync(_buffer.Memory, cancellationToken).ConfigureAwait(false);
+            if (size == 0)
+            {
+                throw new InvalidOperationException("Should never happen: stream is empty?");
+            }
+            using var response = await SendChunkAsync(size, final, cancellationToken).ConfigureAwait(false);
+            if (final)
+            {
                 if (response.StatusCode != HttpStatusCode.OK && response.StatusCode != HttpStatusCode.Created)
                 {
                     throw new GoogleCloudStorageUploadException($"Upload final chunk failed with status code {response.StatusCode}.");
                 }
+                // final chunk upload successfull
+                Sent += size;
+                return;
+            }
+            if (response.StatusCode != /* 308 */ HttpStatusCode.PermanentRedirect)
+            {
+                throw new GoogleCloudStorageUploadException($"Upload chunk failed with status code {response.StatusCode}.");
+            }
+            // chunk upload successfull (TODO: check if response range is set properly)
+            Sent += size;
+        }
+    }
+
+    [Obsolete("Use UploadAsync instead")]
+    public async Task ConumeStreamAsync(Stream stream, CancellationToken cancellationToken)
+    {
+        ThrowIfDisposed();
+        var read = await FillBuffer(stream, cancellationToken).ConfigureAwait(false);
+        if (read == _buffer.Memory.Length)
+        {
+            // partial upload
+            var response = await SendChunkAsync(read, false, cancellationToken).ConfigureAwait(false);
+            if ((int)response.StatusCode == 308)
+            {
+                // chunk upload successfull (TODO: check if response range is set properly)
                 Sent += read;
+                await ConumeStreamAsync(stream, cancellationToken).ConfigureAwait(false);
             }
-        }
-
-        public void Dispose()
-        {
-            if (0 == Interlocked.CompareExchange(ref _isDisposed, 1, 0))
+            else
             {
-                _buffer.Dispose();
-                Client.Dispose();
+                throw new GoogleCloudStorageUploadException($"Upload chunk failed with status code {response.StatusCode}.");
             }
         }
-
-        public ValueTask DisposeAsync()
+        else
         {
-            if (0 == Interlocked.CompareExchange(ref _isDisposed, 1, 0))
+            // FIXME: handle stream being divisable by chunk size
+            var response = await SendChunkAsync(read, true, cancellationToken).ConfigureAwait(false);
+            if (response.StatusCode != HttpStatusCode.OK && response.StatusCode != HttpStatusCode.Created)
             {
-                _buffer.Dispose();
-                Client.Dispose();
+                throw new GoogleCloudStorageUploadException($"Upload final chunk failed with status code {response.StatusCode}.");
             }
-            return default;
+            Sent += read;
         }
+    }
+
+    public void Dispose()
+    {
+        if (0 == Interlocked.CompareExchange(ref _isDisposed, 1, 0))
+        {
+            _buffer.Dispose();
+            Client.Dispose();
+        }
+    }
+
+    public ValueTask DisposeAsync()
+    {
+        if (0 == Interlocked.CompareExchange(ref _isDisposed, 1, 0))
+        {
+            _buffer.Dispose();
+            Client.Dispose();
+        }
+        return default;
     }
 }
